requirements.txt
requests==2.23.0
merry==0.2.2
celery
redis
gevent==20.9.0
kafka-python==2.0.2
loguru
lxml
peewee
pymysql
fastapi
pydantic
uvicorn
jieba
openpyxl
matplotlib
pandas
numpy
wordcloud
Pillow